refactor: move FatalErrorStateModal component
